Update straggling HashMap to LinMultiMap.
When joining NFAs, we keep a map from the other NFA's states to this
NFA's states. This commit switches the data structure from a HashMap to
a LinMultiMap, for a 1.69x speedup!

Digression: LinMultiMap is slightly inappropriate here -- it should
really be a regular map, but I haven't written a plain LinMap yet.

    name                              bench.with_linset ns/iter  bench.with_linmultimap_statemap ns/iter  diff ns/iter   diff %  speedup
    regex::tests::bench_regex_to_nfa  624,438                    370,077                                      -254,361  -40.73%   x 1.69